adicionar carregar nos tickets e noas queue no dashboard

+++ app/Http/Controllers/FullCalenderController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Gate;
 use App\Models\Event;
 
 class FullCalenderController extends Controller
@@ -21,38 +21,41 @@ class FullCalenderController extends Controller
     }
 
     public function action(Request $request)
-    {
-    	if($request->ajax())
-    	{
-    		if($request->type == 'add')
-    		{
-    			$event = Event::create([
-    				'title'		=>	$request->title,
-    				'start'		=>	$request->start,
-    				'end'		=>	$request->end
-    			]);
+    {	
+		if(Gate::allows('is-admin')){
+        
+			if($request->ajax())
+			{
+				if($request->type == 'add')
+				{
+					$event = Event::create([
+						'title'		=>	$request->title,
+						'start'		=>	$request->start,
+						'end'		=>	$request->end
+					]);
 
-    			return response()->json($event);
-    		}
+					return response()->json($event);
+				}
 
-    		if($request->type == 'update')
-    		{
-    			$event = Event::find($request->id)->update([
-    				'title'		=>	$request->title,
-    				'start'		=>	$request->start,
-    				'end'		=>	$request->end
-    			]);
+				if($request->type == 'update')
+				{
+					$event = Event::find($request->id)->update([
+						'title'		=>	$request->title,
+						'start'		=>	$request->start,
+						'end'		=>	$request->end
+					]);
 
-    			return response()->json($event);
-    		}
+					return response()->json($event);
+				}
 
-    		if($request->type == 'delete')
-    		{
-    			$event = Event::find($request->id)->delete();
+				if($request->type == 'delete')
+				{
+					$event = Event::find($request->id)->delete();
 
-    			return response()->json($event);
-    		}
-    	}
+					return response()->json($event);
+				}
+			}
+		}
     }
 }
 ?>
